Updates to be able to use next track, prior track and resume
Since the source of audio is HTTP and HTTP Close occurs anyway on a Pause, The Media Player is just doing a stop while maintaining playlist information.  On resume, the pipeline is started, but the current URL is used on the restart.  From a listening perspective, the song is restarted on resume, but this was happening anyway (with the pipeline spitting out some of the song prior to restarting).  Next Track and Prior Track require customization of the ESPHome Integration MediaPlayer to enable.